Global and Local rotational energies supported
Now there are two different Rotational Inertia energies, one that uses
derivatives with respect to a local angle, and one that uses derivatives
with respect to the global angle.
This feature can be supported because now the inertia energies are the
ones responsible to update the simulation state after each Newton Step.