First differentiable rods
The initialization of the Rod has been changed so that the rod segments
have first the RB_A and then the RB_B.
Fixed a bug in compose axis angle where because of a threshold too low
it divided by zero.
Introduced a new way of computing the darboux derivative.
Also fixed a bug in the obj curve obj loader.